Support overlays in iframes
It's currently tricky to use an overlay inside of an iframe.

Since most ACT overlays load files from disk, file urls are treated
as cross-domain.  This means that events can't be posted directly
to the iframe, and have to be posted via a post message.

Update all of the overlay themes to support messages in addition to the
custom events.  The 'type' field on the message is treated as equivalent
to the event name, so that multiple messages can be handled.